test functionl need to implement for all
