+++ mobile/package.json
@@ -21,6 +21,7 @@
     "@types/react-native-elements": "^0.18.0",
     "@types/react-native-material-ui": "^1.32.3",
     "expo": "^48.0.19",
+    "expo-app-loading": "^2.1.1",
     "expo-status-bar": "~1.4.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
